pyproject.toml
[project]
name = "langchain-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[anthropic]>=1.0.5",
    "python-dotenv>=1.2.1",
]
